入力検証とデッドロック対策を追加する
- コールバック実行時のデッドロックを修正する
- sample_rate の正値検証を追加する
- max_video_queue_size の負値検証を追加する
- I420/NV12 の偶数サイズ検証を追加する
- audio_player.cpp に <algorithm> ヘッダを追加する

+++ src/bindings/video_player.h
@@ -113,7 +113,7 @@ class VideoPlayer {
   nb::dict stats() const;
 
   // === キューサイズ制御 ===
-  void set_max_video_queue_size(size_t size);
+  void set_max_video_queue_size(int64_t size);
   size_t get_max_video_queue_size() const;
   void drain_video();
